+++ cmd/effect/internal/effect4/templates/models.ts.gotmpl
