+++ cmd/effect/internal/effect4/templates/repository.ts.gotmpl
@@ -1,0 +1,61 @@
+{{define "service-tag" -}}
+// Service Tag
+export class {{.RepositoryName}} extends Context.Service<{{.RepositoryName}}>()("{{.RepositoryName}}", { make: {{.RepositoryNameCamel}}Make }) {
+  static readonly layer = Layer.effect(this, this.make)
+}
+
+export type {{.RepositoryName}}Service = {{.RepositoryName}}["Service"]
+{{- end}}
+
+{{define "execute-fn" -}}
+{{- if .UseTemplateLiterals -}}
+{{- if .HasParams}}(params) => {{else}}() => {{end -}}sql`{{.SQLTemplateLiteral}}`
+{{- else -}}
+{{- if .HasParams}}(params) => {{else}}() => {{end -}}sql.unsafe(`{{.OriginalSQL}}`{{if .HasParams}}, [{{.ParamList}}]{{end}})
+{{- end -}}
+{{- end}}
+
+{{define "query-impl"}}
+  // {{.Name}}
+{{- $sqlLines := splitLines .OriginalSQL }}
+{{- range $sqlLines }}
+  // {{ . }}
+{{- end }}
+  const {{.NameCamel}} = {{.SqlSchemaMethod}}({
+    Request: {{.RequestSchema}},
+{{- if .HasResults }}
+    Result: {{.NamePascal}}Result,
+{{- end }}
+    execute: {{template "execute-fn" .}}
+  })
+{{- end}}
+
+{{define "implementation" -}}
+// Implementation
+const {{.RepositoryNameCamel}}Make = Effect.gen(function* () {
+  const sql = yield* SqlClient.SqlClient
+{{range .QueryViews}}
+{{template "query-impl" .}}
+{{- end}}
+
+  return {
+{{- range .QueryViews }}
+    {{.NameCamel}},
+{{- end }}
+  } as const
+})
+{{- end}}
+
+{{define "live-layer" -}}
+// Live Layer
+export const {{.RepositoryNameCamel}}Live = {{.RepositoryName}}.layer
+{{- end}}
+
+// Generated by sqlc {{.SqlcVersion}}, plugin: sqlc-effect {{.PluginVersion}}
+{{formatImports .Imports}}
+
+{{template "implementation" .}}
+
+{{template "service-tag" .}}
+
+{{template "live-layer" .}}
