+++ cmd/effect/internal/effect4/templates/response.ts.gotmpl
