+++ cmd/effect/internal/version/version.go
@@ -1,0 +1,5 @@
+package version
+
+// Version is the version of this sqlc-gen-effect plugin.
+// This is set at build time via ldflags by goreleaser.
+var Version = "dev"
