--- cmd/plugin/internal/version/version.go
@@ -1,5 +1,0 @@
-package version
-
-// Version is the version of this sqlc-gen-better-typescript plugin.
-// This is set at build time via ldflags by goreleaser.
-var Version = "dev"
