+++ examples/effect-v4/sqlc.yaml
@@ -11,7 +11,6 @@ sql:
   - out: src/repositories
     plugin: effect
     options: 
-      builder: effect-v4-unstable
       import_extension: ".js"
       # disable_template_literals: true
       # debug: true
