+++ toolbelt/file.go
@@ -1,0 +1,7 @@
+package toolbelt
+
+// File represents a generated file returned to sqlc.
+type File struct {
+	Name    string
+	Content []byte
+}
